+++ CoShift/src/main/java/org/coshift/c_adapters/dto/ShiftDto.java
@@ -5,10 +5,10 @@ import java.util.List;
 
 
 public record ShiftDto(
-    Long   id,
-    LocalDateTime startTime,          // ISO-8601, z. B. „2025-06-22T14:00“
-    int    durationInMinutes,
+    long   id,
+    LocalDateTime startTime,
+    long durationInMinutes,
     int    capacity,
-    List<Long> personIds                  // geändert: nur noch IDs statt PersonDto
+    List<Long> personIds
 ) 
 {}
